Replace GaussianMLPPolicy and Baseline
