+++ tests/garage/tf/algos/test_tnpg.py
@@ -34,6 +34,6 @@ class TestTNPG(TfGraphTestCase):
             runner.setup(algo, env)
 
             last_avg_ret = runner.train(n_epochs=10, batch_size=10000)
-            assert last_avg_ret > 30
+            assert last_avg_ret > 15
 
             env.close()
