+++ tests/garage/tf/policies/test_gaussian_mlp_policy.py
@@ -6,13 +6,14 @@ import pytest
 import tensorflow as tf
 
 from garage.tf.envs import TfEnv
-from garage.tf.policies import GaussianMLPPolicyWithModel
+from garage.tf.policies import GaussianMLPPolicy
 from tests.fixtures import TfGraphTestCase
 from tests.fixtures.envs.dummy import DummyBoxEnv
 from tests.fixtures.models import SimpleGaussianMLPModel
 
 
-class TestGaussianMLPPolicyWithModel(TfGraphTestCase):
+class TestGaussianMLPPolicy(TfGraphTestCase):
+
     @pytest.mark.parametrize('obs_dim, action_dim', [
         ((1, ), (1, )),
         ((1, ), (2, )),
@@ -21,12 +22,14 @@ class TestGaussianMLPPolicyWithModel(TfGraphTestCase):
         ((1, 1), (2, 2)),
         ((2, 2), (2, 2)),
     ])
-    def test_get_action(self, obs_dim, action_dim):
+    @mock.patch('numpy.random.normal')
+    def test_get_action(self, mock_normal, obs_dim, action_dim):
+        mock_normal.return_value = 0.5
         env = TfEnv(DummyBoxEnv(obs_dim=obs_dim, action_dim=action_dim))
         with mock.patch(('garage.tf.policies.'
-                         'gaussian_mlp_policy_with_model.GaussianMLPModel'),
+                         'gaussian_mlp_policy.GaussianMLPModel'),
                         new=SimpleGaussianMLPModel):
-            policy = GaussianMLPPolicyWithModel(env_spec=env.spec)
+            policy = GaussianMLPPolicy(env_spec=env.spec)
 
         env.reset()
         obs, _, _, _ = env.step(1)
@@ -35,7 +38,7 @@ class TestGaussianMLPPolicyWithModel(TfGraphTestCase):
 
         expected_action = np.full(action_dim, 0.75)
         expected_mean = np.full(action_dim, 0.5)
-        expected_log_std = np.full(action_dim, 0.5)
+        expected_log_std = np.full(action_dim, np.log(0.5))
 
         assert env.action_space.contains(action)
         assert np.array_equal(action, expected_action)
@@ -61,9 +64,9 @@ class TestGaussianMLPPolicyWithModel(TfGraphTestCase):
     def test_dist_info_sym(self, obs_dim, action_dim):
         env = TfEnv(DummyBoxEnv(obs_dim=obs_dim, action_dim=action_dim))
         with mock.patch(('garage.tf.policies.'
-                         'gaussian_mlp_policy_with_model.GaussianMLPModel'),
+                         'gaussian_mlp_policy.GaussianMLPModel'),
                         new=SimpleGaussianMLPModel):
-            policy = GaussianMLPPolicyWithModel(env_spec=env.spec)
+            policy = GaussianMLPPolicy(env_spec=env.spec)
 
         env.reset()
         obs, _, _, _ = env.step(1)
@@ -75,7 +78,7 @@ class TestGaussianMLPPolicyWithModel(TfGraphTestCase):
 
         # flatten output
         expected_mean = [np.full(np.prod(action_dim), 0.5)]
-        expected_log_std = [np.full(np.prod(action_dim), 0.5)]
+        expected_log_std = [np.full(np.prod(action_dim), np.log(0.5))]
 
         prob = self.sess.run(dist1_sym, feed_dict={obs_ph: [obs.flatten()]})
 
@@ -93,16 +96,16 @@ class TestGaussianMLPPolicyWithModel(TfGraphTestCase):
     def test_is_pickleable(self, obs_dim, action_dim):
         env = TfEnv(DummyBoxEnv(obs_dim=obs_dim, action_dim=action_dim))
         with mock.patch(('garage.tf.policies.'
-                         'gaussian_mlp_policy_with_model.GaussianMLPModel'),
+                         'gaussian_mlp_policy.GaussianMLPModel'),
                         new=SimpleGaussianMLPModel):
-            policy = GaussianMLPPolicyWithModel(env_spec=env.spec)
+            policy = GaussianMLPPolicy(env_spec=env.spec)
 
         env.reset()
         obs, _, _, _ = env.step(1)
         obs_dim = env.spec.observation_space.flat_dim
 
-        with tf.compat.v1.variable_scope(
-                'GaussianMLPPolicyWithModel/GaussianMLPModel', reuse=True):
+        with tf.compat.v1.variable_scope('GaussianMLPPolicy/GaussianMLPModel',
+                                         reuse=True):
             return_var = tf.compat.v1.get_variable('return_var')
         # assign it to all one
         return_var.load(tf.ones_like(return_var).eval())
